+++ Encoding/NVEnc.vb
@@ -556,6 +556,16 @@ Public Class NVEnc
         End Sub
 
         Protected Overrides Sub OnValueChanged(item As CommandLineParam)
+            If Not Decoder.MenuButton Is Nothing AndAlso item Is Decoder OrElse item Is Nothing Then
+                Dim isIntelPresent = OS.VideoControllers.Where(Function(val) val.Contains("AMD") OrElse val.Contains("NVIDIA")).Count <> OS.VideoControllers.Length
+
+                For x = 0 To Decoder.Options.Length - 1
+                    If Decoder.Options(x).Contains("Intel") Then
+                        Decoder.ShowOption(x, isIntelPresent)
+                    End If
+                Next
+            End If
+
             If Not QPI.NumEdit Is Nothing Then
                 VppNnediField.MenuButton.Enabled = VppNnedi.Value
                 VppNnediNns.MenuButton.Enabled = VppNnedi.Value