+++ changelog.md
@@ -6,6 +6,8 @@
   with an underscore and thus being lost, now the characters
   are escaped/unescaped and thus preserved
 - fix: update x265 switch --refine-mv
+- fix: in the nvenc dialog, Intel decoding options were present
+  even when no Intel GPU is present
 - new: option to check online for new stable version once per day
 
 ### 2.0.4.3 Beta